Refactor attack effect snapshot utilities
